split ut
